EVO: format code (if/return)

+++ src/MonarcFO/Service/AnrRecommandationRiskService.php
@@ -155,36 +155,32 @@ class AnrRecommandationRiskService extends \MonarcCore\Service\AbstractService
             foreach ($recommandationsRisks as $recommandationRisk) {
                 if ($recommandationRisk->recommandation->id == $recommandation->id) {
                     //retrieve instance risk associated
-                    if ($recommandationRisk->instanceRisk) {
-                        if ($recommandationRisk->instanceRisk->kindOfMeasure != InstanceRisk::KIND_NOT_TREATED) {
-                            $instanceRisk = $recommandationRisk->instanceRisk;
-                            if (is_object($instanceRisk->asset)) {
-                                $instanceRisk->asset = $instanceRisk->asset->getJsonArray();
-                            }
-                            if (is_object($instanceRisk->threat)) {
-                                $instanceRisk->threat = $instanceRisk->threat->getJsonArray();
-                            }
-                            if (is_object($instanceRisk->vulnerability)) {
-                                $instanceRisk->vulnerability = $instanceRisk->vulnerability->getJsonArray();
-                            }
-                            $riskData = $instanceRisk->getJsonArray();
-                            $riskData['instance'] = $this->instanceTable->getEntity($riskData['instance'])->getJsonArray();
-                            $recommandations[$key]['risks'][] = $riskData;
-                            $nbRisks++;
+                    if ($recommandationRisk->instanceRisk && $recommandationRisk->instanceRisk->kindOfMeasure != InstanceRisk::KIND_NOT_TREATED) {
+                        $instanceRisk = $recommandationRisk->instanceRisk;
+                        if (is_object($instanceRisk->asset)) {
+                            $instanceRisk->asset = $instanceRisk->asset->getJsonArray();
                         }
+                        if (is_object($instanceRisk->threat)) {
+                            $instanceRisk->threat = $instanceRisk->threat->getJsonArray();
+                        }
+                        if (is_object($instanceRisk->vulnerability)) {
+                            $instanceRisk->vulnerability = $instanceRisk->vulnerability->getJsonArray();
+                        }
+                        $riskData = $instanceRisk->getJsonArray();
+                        $riskData['instance'] = $this->instanceTable->getEntity($riskData['instance'])->getJsonArray();
+                        $recommandations[$key]['risks'][] = $riskData;
+                        $nbRisks++;
                     }
                     //retrieve instance risk op associated
-                    if ($recommandationRisk->instanceRiskOp) {
-                        if ($recommandationRisk->instanceRiskOp->kindOfMeasure != InstanceRiskOp::KIND_NOT_TREATED) {
-                            $data = $recommandationRisk->instanceRiskOp->getJsonArray();
-                            $instance = $recommandationRisk->instanceRiskOp->instance->getJsonArray();
-                            unset($instance['__initializer__']);
-                            unset($instance['__cloner__']);
-                            unset($instance['__isInitialized__']);
-                            $data['instance'] = $instance;
-                            $recommandations[$key]['risksop'][] = $data;
-                            $nbRisks++;
-                        }
+                    if ($recommandationRisk->instanceRiskOp && $recommandationRisk->instanceRiskOp->kindOfMeasure != InstanceRiskOp::KIND_NOT_TREATED) {
+                        $data = $recommandationRisk->instanceRiskOp->getJsonArray();
+                        $instance = $recommandationRisk->instanceRiskOp->instance->getJsonArray();
+                        unset($instance['__initializer__']);
+                        unset($instance['__cloner__']);
+                        unset($instance['__isInitialized__']);
+                        $data['instance'] = $instance;
+                        $recommandations[$key]['risksop'][] = $data;
+                        $nbRisks++;
                     }
 
                     //delete risk of global with risk value is not the higher
@@ -418,7 +414,7 @@ class AnrRecommandationRiskService extends \MonarcCore\Service\AbstractService
 
         //verify if risk is final or intermediate (risk attach to others recommandations)
         $riskRecommandations = $table->getEntityByFields(['instanceRisk' => $recoRisk->instanceRisk->id]);
-        $final = (count($riskRecommandations) == 1) ? true : false;
+        $final = (count($riskRecommandations) == 1);
 
         //automatically record in history before modify recommendation and risk valuesc
         $this->createRecoHistoric($data, $recoRisk, $final);
@@ -697,10 +693,8 @@ class AnrRecommandationRiskService extends \MonarcCore\Service\AbstractService
 
                 $recommandationRisksReco = $table->getEntityByFields(['anr' => $recommandationRisk->anr->id, 'recommandation' => $recommandationRisk->recommandation->id]);
                 foreach ($recommandationRisksReco as $recommandationRiskReco) {
-                    if ($recommandationRiskReco->instanceRisk) {
-                        if (in_array($recommandationRiskReco->instanceRisk->id, $brothersIds)) {
-                            $this->get('table')->delete($recommandationRiskReco->id);
-                        }
+                    if ($recommandationRiskReco->instanceRisk && in_array($recommandationRiskReco->instanceRisk->id, $brothersIds)) {
+                        $this->get('table')->delete($recommandationRiskReco->id);
                     }
                 }
             } else {
@@ -720,10 +714,8 @@ class AnrRecommandationRiskService extends \MonarcCore\Service\AbstractService
 
                 $recommandationRisksReco = $table->getEntityByFields(['anr' => $recommandationRisk->anr->id, 'recommandation' => $recommandationRisk->recommandation->id]);
                 foreach ($recommandationRisksReco as $recommandationRiskReco) {
-                    if ($recommandationRiskReco->instanceRiskOp) {
-                        if (in_array($recommandationRiskReco->instanceRiskOp->id, $brothersIds)) {
-                            $this->get('table')->delete($recommandationRiskReco->id);
-                        }
+                    if ($recommandationRiskReco->instanceRiskOp && in_array($recommandationRiskReco->instanceRiskOp->id, $brothersIds)) {
+                        $this->get('table')->delete($recommandationRiskReco->id);
                     }
                 }
             } else {
